requirements.txt
autopep8==1.5.7
backcall==0.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
debugpy==1.3.0
decorator==5.0.9
fastapi==0.65.3
ffmpeg==1.4
flake8==3.9.2
greenlet==1.1.0
h11==0.12.0
idna==2.10
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
lxml==4.6.3
matplotlib-inline==0.1.2
mccabe==0.6.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycodestyle==2.7.0
pydantic==1.8.2
pyflakes==2.3.1
Pygments==2.9.0
python-dateutil==2.8.1
python-dotenv==0.18.0
pyzmq==22.1.0
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.20
SQLAlchemy-Utils==0.37.8
starlette==0.14.2
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
wcwidth==0.2.5
youtube-dl==2021.6.6
ffprobe-python